requirements.txt
requests
python-dotenv
pydantic
jupyter==1.0.0
numpy
pandas
scipy
matplotlib
seaborn
scikit-learn
tqdm
sqlalchemy
colorama